+++ widgets/telemetry-graphs/api.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  async getTelemetry({ homey, query }) {
+    return homey.app.getTelemetryWidgetState(query.deviceId);
+  },
+};